fixed the verifiers

+++ backend/schedule/ScheduleController.cs
@@ -28,27 +28,13 @@ namespace capstone
             }
             public async Task<bool> VerifyAccount(string? Id) {
                 if(string.IsNullOrEmpty(Id)) return false;
-                Account temp;
-                try {
-                    temp = accounts.Find(user => user._id == ObjectId.Parse(Id)).ToList().First();
-                }
-                catch(Exception _i) {
-                    return false;
-                }
-                return true;
+
+                return accounts.Find(user => user._id == ObjectId.Parse(Id)).ToList().Any();
             }
             public async Task<bool> VerifyCourse(string? Id) {
                 if (string.IsNullOrEmpty(Id)) return false;
-                Course temp;
-                try
-                {
-                    temp = courses.Find(course => course._id == ObjectId.Parse(Id)).ToList().First();
-                }
-                catch (Exception _i)
-                {
-                    return false;
-                }
-                return true;
+
+                return courses.Find(course => course._id == ObjectId.Parse(Id)).ToList().Any();
             }
             private async Task<bool> CheckAuthorization(string? auth)
             {
@@ -98,13 +84,7 @@ namespace capstone
             public async Task<IResult> DeleteSchedule(string id, string? auth) {
                 if (!await CheckAuthorization(auth)) return Results.BadRequest("Invalid authorization");
 
-                Schedule schedule;
-                try {
-                    schedule = schedules.Find(schedule => schedule._id == ObjectId.Parse(id)).ToList().First();
-                }
-                catch(Exception _i) {
-                    return Results.BadRequest("Schedule id is invalid");
-                }
+                if(!schedules.Find(schedule => schedule._id == ObjectId.Parse(id)).ToList().Any()) return Results.BadRequest("Schedule id is invalid");
 
                 var filter = Builders<Schedule>.Filter.Eq(s => s._id, ObjectId.Parse(id));
                 await schedules.DeleteOneAsync(filter);
